Сортировка правки
